Crate bridge_hub_test_utils

Module contains predefined test-case scenarios for “BridgeHub” Runtimes.

Modules

test_cases

Module contains predefined test-case scenarios for Runtime with bridging capabilities.

test_data

Generating test data, used by all tests.

Structs

CollatorSessionKey

CollatorSessionKeys

ExtBuilder

Basic builder based on balances, collators and pallet_session.

RuntimeHelper

SlotDurations

Enums

XcmReceivedFrom

Traits

BasicParachainRuntime

A set of traits for a minimal parachain runtime, that may be used in conjunction with the ExtBuilder and the RuntimeHelper.

Functions

assert_metadata

assert_total

check_sane_fees_values

A helper function for comparing the actual value of a fee constant with its estimated value. The estimated value can be overestimated (overestimate_in_percent), and if the difference to the actual value is below margin_overestimate_diff_in_percent_for_lowering, we should lower the actual value.

diff_as_percent

mock_open_hrmp_channel

Helper function which emulates opening HRMP channel which is needed for XcmpQueue to pass.

test_header

Get a header for testing.

Type Aliases

AccountIdOf

BalanceOf

RuntimeCallOf

SessionKeysOf

ValidatorIdOf