Function bridge_hub_test_utils::test_cases::handle_export_message_from_system_parachain_to_outbound_queue_works

pub fn handle_export_message_from_system_parachain_to_outbound_queue_works<Runtime, XcmConfig, MessagesPalletInstance>(
    collator_session_key: CollatorSessionKeys<Runtime>,
    runtime_para_id: u32,
    sibling_parachain_id: u32,
    unwrap_pallet_bridge_messages_event: Box<dyn Fn(Vec<u8>) -> Option<Event<Runtime, MessagesPalletInstance>>>,
    export_message_instruction: fn() -> Instruction<XcmConfig::RuntimeCall>,
    expected_lane_id: LaneId
)where
    Runtime: Config + Config + Config + Config + Config + Config + Config + Config + Config<MessagesPalletInstance>,
    XcmConfig: Config,
    MessagesPalletInstance: 'static,
    ValidatorIdOf<Runtime>: From<AccountIdOf<Runtime>>,

Test-case makes sure that Runtime can handle xcm ExportMessage: Checks if received XCM messages is correctly added to the message outbound queue for delivery. For SystemParachains we expect unpaid execution.