Type Alias bridge_hub_test_utils::test_cases::RuntimeHelper

pub type RuntimeHelper<Runtime, AllPalletsWithoutSystem = ()> = RuntimeHelper<Runtime, AllPalletsWithoutSystem>;

Aliased Type

struct RuntimeHelper<Runtime, AllPalletsWithoutSystem = ()>(/* private fields */);

Implementations

impl<Runtime, AllPalletsWithoutSystem> RuntimeHelper<Runtime, AllPalletsWithoutSystem>where Runtime: Config, <Runtime as Config>::AccountId: Into<<<Runtime as Config>::RuntimeOrigin as OriginTrait>::AccountId>, AllPalletsWithoutSystem: OnInitialize<<<<Runtime as Config>::Block as HeaderProvider>::HeaderT as Header>::Number> + OnFinalize<<<<Runtime as Config>::Block as HeaderProvider>::HeaderT as Header>::Number>,

Utility function that advances the chain to the desired block number. If an author is provided, that author information is injected to all the blocks in the meantime.

pub fn run_to_block( n: u32, author: <Runtime as Config>::AccountId ) -> <<Runtime as Config>::Block as HeaderProvider>::HeaderT

pub fn root_origin() -> <Runtime as Config>::RuntimeOrigin

pub fn origin_of( account_id: <Runtime as Config>::AccountId ) -> <Runtime as Config>::RuntimeOrigin

impl<XcmConfig, AllPalletsWithoutSystem> RuntimeHelper<XcmConfig, AllPalletsWithoutSystem>where XcmConfig: Config,

pub fn do_transfer( from: MultiLocation, to: MultiLocation, _: (MultiLocation, u128) ) -> Result<Assets, Error>

impl<Runtime, AllPalletsWithoutSystem> RuntimeHelper<Runtime, AllPalletsWithoutSystem>where Runtime: Config + Config,

pub fn do_teleport_assets<HrmpChannelOpener>( origin: <Runtime as Config>::RuntimeOrigin, dest: MultiLocation, beneficiary: MultiLocation, _: (MultiLocation, u128), open_hrmp_channel: Option<(u32, u32)>, included_head: <<Runtime as Config>::Block as HeaderProvider>::HeaderT, slot_digest: &[u8] ) -> Result<(), DispatchError>where HrmpChannelOpener: ProvideInherent<Call = Call<Runtime>>,

impl<Runtime, AllPalletsWithoutSystem> RuntimeHelper<Runtime, AllPalletsWithoutSystem>where Runtime: Config + Config,

pub fn execute_as_governance( call: Vec<u8, Global>, require_weight_at_most: Weight ) -> Outcome

impl<ParachainSystem, AllPalletsWithoutSystem> RuntimeHelper<ParachainSystem, AllPalletsWithoutSystem>where ParachainSystem: Config,

pub fn xcm_max_weight(from: XcmReceivedFrom) -> Weight

impl<Runtime, AllPalletsWithoutSystem> RuntimeHelper<Runtime, AllPalletsWithoutSystem>where Runtime: Config + Config,

pub fn assert_pallet_xcm_event_outcome( unwrap_pallet_xcm_event: &Box<dyn Fn(Vec<u8, Global>) -> Option<Event<Runtime>>, Global>, assert_outcome: fn(_: Outcome) )

pub fn get_pallet_xcm_event_outcome( unwrap_pallet_xcm_event: &Box<dyn Fn(Vec<u8, Global>) -> Option<Event<Runtime>>, Global> ) -> Outcome

impl<Runtime, AllPalletsWithoutSystem> RuntimeHelper<Runtime, AllPalletsWithoutSystem>where Runtime: Config + Config,

pub fn xcmp_queue_message_sent( unwrap_xcmp_queue_event: Box<dyn Fn(Vec<u8, Global>) -> Option<Event<Runtime>>, Global> ) -> Option<[u8; 32]>

impl<HrmpChannelSource, AllPalletsWithoutSystem> RuntimeHelper<HrmpChannelSource, AllPalletsWithoutSystem>where HrmpChannelSource: XcmpMessageSource,

pub fn take_xcm(sent_to_para_id: Id) -> Option<VersionedXcm<()>>