Module bridge_hub_test_utils::test_data::from_parachain

Generating test data for bridges with remote parachains.

Functions

make_complex_bridged_parachain_heads_proof

Make bridged parachain header with given state root and relay header that is finalizing it.

make_complex_relayer_confirmation_batch

Prepare a batch call with relay finality proof, parachain head proof and message delivery proof.

make_complex_relayer_confirmation_proofs

Prepare storage proofs of message confirmations, stored at the target parachain.

make_complex_relayer_delivery_batch

Prepare a batch call with relay finality proof, parachain head proof and message proof.

make_complex_relayer_delivery_proofs

Prepare storage proofs of messages, stored at the source chain.