Module bridge_hub_test_utils::test_cases

Module contains predefined test-case scenarios for Runtime with bridging capabilities.

This file contains tests, suitable for all bridge runtimes. See from_parachain and from_grandpa_chain submodules for tests, that are specific to the bridged chain type.

Modules

from_grandpa_chain

Module contains predefined test-case scenarios for Runtime with bridging capabilities with remote GRANDPA chain.

from_parachain

Module contains predefined test-case scenarios for Runtime with bridging capabilities with remote parachain.

Macros

include_teleports_for_native_asset_works

Functions

can_calculate_weight_for_paid_export_message_with_reserve_transfer

Estimates XCM execution fee for paid ExportMessage processing.

change_bridge_grandpa_pallet_mode_by_governance_works

Test-case makes sure that Runtime can change bridge GRANDPA pallet operating mode via governance-like call.

change_bridge_messages_pallet_mode_by_governance_works

Test-case makes sure that Runtime can change bridge messaging pallet operating mode via governance-like call.

change_bridge_parachains_pallet_mode_by_governance_works

Test-case makes sure that Runtime can change bridge parachains pallet operating mode via governance-like call.

change_storage_constant_by_governance_works

Test-case makes sure that Runtime can change storage constant via governance-like call

handle_export_message_from_system_parachain_to_outbound_queue_works

Test-case makes sure that Runtime can handle xcm ExportMessage: Checks if received XCM messages is correctly added to the message outbound queue for delivery. For SystemParachains we expect unpaid execution.

initialize_bridge_by_governance_works

Test-case makes sure that Runtime can process bridging initialize via governance-like call

message_dispatch_routing_works

Test-case makes sure that Runtime can route XCM messages received in inbound queue, We just test here MessageDispatch configuration. We expect that runtime can route messages: 1. to Parent (relay chain) 2. to Sibling parachain

open_and_close_bridge_works

Test-case makes sure that Runtime can open/close bridges.

run_test

Prepare default runtime storage and run test within this context.

set_storage_keys_by_governance_works

Test-case makes sure that Runtime can change storage constant via governance-like call

Type Aliases

RuntimeHelper