Module bridge_hub_test_utils::test_cases

Module contains predefined test-case scenarios for Runtime with bridging capabilities.

Macros

include_teleports_for_native_asset_works

Structs

CollatorSessionKeys

Enums

XcmReceivedFrom

Functions

handle_export_message_from_system_parachain_to_outbound_queue_works

Test-case makes sure that Runtime can handle xcm ExportMessage: Checks if received XCM messages is correctly added to the message outbound queue for delivery. For SystemParachains we expect unpaid execution.

initialize_bridge_by_governance_works

Test-case makes sure that Runtime can process bridging initialize via governance-like call

message_dispatch_routing_works

Test-case makes sure that Runtime can route XCM messages received in inbound queue, We just test here MessageDispatch configuration. We expect that runtime can route messages: 1. to Parent (relay chain) 2. to Sibling parachain